fix: update imports to use types from v6 instead of v5
